\name{lcms3}
\alias{lcms3}
\docType{data}
\title{
	a real MS dataset obtained from analyzing human sera
}
\description{
	This data set consists of three-replicated LC/MS/MS data obtained from laboratory of gaseous ion chemistry in department of chemistry, Korea University. 
}
\usage{data(lcms3)}
\details{
	LC/MS/MS experiments were performed on a home-built ultrahigh-pressure dual on-line solid phase extraction/capillary reverse-phase liquid chromatography (DO-SPE/cRPLC) (Min et al., 2007) that was coupled to a Fourier transform ion cyclotron resonance (FT-ICR, LTQ-FT, Thermo, San Jose) mass spectrometer by means of a home built nanoESI interface.
}
\format{
	a matrix for LCMS data, rows = peptides, columns = samples
}
\source{
	Min H-K, Hyung S-W, Shin J-W, et al. (2007). Ultrahigh-pressure dual online solid phase extraction/capillary reverse-phase liquid chromatography/tandem mass spectrometry (DO-SPE/cRPLC/MS/MS): A versatile separation platform for high-throughput and highly sensitive proteomic analyses. Electrophoresis 28:1012–1021. 
}
\keyword{datasets}

