\name{lcms3}
\alias{lcms3}
\docType{data}
\title{
	LC/MS/MS data
}
\description{
  This data set consists of three-replicated LC/MS/MS data obtained from laboratory of gaseous ion chemistry in department of chemistry, Korea University. The data is already converted into log-2 transformation.
}
\usage{data(lcms3)}
\format{a matrix for LCMS data, rows=peptides, columns=samples}
\source{
	Min et al. (2007). Ultrahigh-pressure dual online solid phase extraction/capillary reverse-phase liquid chromatography/tandem mass spectrometry (DO-SPE/cRPLC/MS/MS): A versatile separation platform for high-throughput and highly sensitive proteomic analyses. \emph{ELECTROPHORESIS}, \bold{28}, 1012--1021. 
}
\examples{
data(lcms3)
}
\keyword{datasets}

